+++ packages/components/scroll-area/src/scroll-area-content.tsx
@@ -1,14 +1,16 @@
 "use client";
 
 import { useRenderElement } from "@jamsrui/hooks";
-import { type UIProps } from "@jamsrui/utils";
+
 import { useScrollAreaContext } from "./scroll-area-context";
-import { scrollArea } from "./styles";
+import { scrollAreaVariants } from "./styles";
+
+import type {UIProps} from "@jamsrui/utils";
 
 export const ScrollAreaContent = (props: ScrollAreaContent.Props) => {
   const { render, children, className, ...restProps } = props;
   const { contentRef } = useScrollAreaContext();
-  const { content } = scrollArea();
+  const { content } = scrollAreaVariants();
   const styles = content({ className });
 
   const renderElement = useRenderElement("div", {