+++ packages/components/scroll-area/src/scroll-area-viewport.tsx
@@ -1,9 +1,11 @@
 "use client";
 
 import { useRenderElement } from "@jamsrui/hooks";
+
 import { useScrollAreaContext } from "./scroll-area-context";
-import { scrollArea } from "./styles";
-import { type UIProps } from "@jamsrui/utils";
+import { scrollAreaVariants } from "./styles";
+
+import type {UIProps} from "@jamsrui/utils";
 
 export const ScrollAreaViewport = (props: ScrollAreaViewport.Props) => {
   const { render, children, className, ...restProps } = props;
@@ -11,7 +13,7 @@ export const ScrollAreaViewport = (props: ScrollAreaViewport.Props) => {
   // Get context
   const { viewportRef, handleScroll } = useScrollAreaContext();
 
-  const { viewport } = scrollArea();
+  const { viewport } = scrollAreaVariants();
   const styles = viewport({ className });
 
   const renderElement = useRenderElement("div", {